core/ai-worker/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain ecosystem
langchain==0.0.350
langchain-community==0.0.1
langchain-experimental==0.0.47
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# AI/ML frameworks
openai==1.3.7
anthropic==0.7.8
tiktoken==0.5.2

# Multi-agent frameworks (Week 3)
swarms==5.1.0
crewai==0.22.5
pyautogen==0.2.0

# Vector databases and embeddings
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.1

# HTTP and async
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Redis for caching and queues
redis==5.0.1
celery==5.3.4

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Utilities
python-multipart==0.0.6
jinja2==3.1.2
